refactor: tech debt cleanup — split large files, clarify TODOs, add coordinator tests
- Split agent-network.tsx (1284→264+197+94 lines) into node/edge/context
- Split agents.tsx (1149→284+208+209 lines) into dialogs/reputation/virtual-grid
- Split messages.tsx (992→169+187+243+146+85 lines) into feed/graph/filter/cards/utils
- Clarify 13 TODO stubs in sandbox-cli with coordinator context
- Add 487 lines of coordinator MCP tool tests (35 tests, needs test runner wiring)
- Net: -3187 lines removed, +3006 lines added
